Update bindAndListen server method
It tries to bind to any of the predefined ports. If all of them are al-
ready in use, it aborts its execution. Otherwie, establishes itself in
the available port and listens normally in there.

The list of possible ports are now in the definitions.h, so the Client
also changed a bit to use that global difined vector.